move StructuredToAvroTransformer to common so that both use
