+++ core-plugins/src/main/java/co/cask/hydrator/plugin/batch/sink/S3AvroBatchSink.java
@@ -26,8 +26,8 @@ import co.cask.cdap.etl.api.Emitter;
 import co.cask.cdap.etl.api.batch.BatchRuntimeContext;
 import co.cask.cdap.etl.api.batch.BatchSink;
 import co.cask.cdap.etl.api.batch.BatchSinkContext;
+import co.cask.hydrator.common.StructuredToAvroTransformer;
 import co.cask.hydrator.plugin.common.Properties;
-import co.cask.hydrator.plugin.common.StructuredToAvroTransformer;
 import com.google.common.collect.Maps;
 import org.apache.avro.generic.GenericRecord;
 import org.apache.avro.mapred.AvroKey;
